refactor telemetry for different host names. update mapbox telemetry lib version.
