+++ MapboxVision/src/main/java/com/mapbox/vision/sync/SessionWriterListener.kt
@@ -1,0 +1,13 @@
+package com.mapbox.vision.sync
+
+import com.mapbox.vision.mobile.core.models.VideoClip
+
+internal interface SessionWriterListener {
+
+    fun onSessionStop(
+        clips: Array<VideoClip>,
+        videoPath: String,
+        outputPath: String,
+        sessionStartMillis: Long
+    )
+}
