+++ MapboxVision/src/main/java/com/mapbox/vision/sync/metagenerator/MetaGenerator.kt
@@ -1,0 +1,13 @@
+package com.mapbox.vision.sync.metagenerator
+
+import com.mapbox.vision.mobile.core.models.VideoClip
+import java.util.HashMap
+
+interface MetaGenerator {
+
+    fun generateMeta(
+        videoClipMap: HashMap<String, VideoClip>,
+        saveDirPath: String,
+        startRecordCoreMillis: Long
+    )
+}
