+++ MapboxVision/src/main/java/com/mapbox/vision/sync/session/SessionWriter.kt
@@ -1,0 +1,8 @@
+package com.mapbox.vision.sync.session
+
+internal interface SessionWriter {
+
+    fun start()
+
+    fun stop()
+}
