+++ MapboxVision/src/main/java/com/mapbox/vision/sync/syncmanager/SyncManager.kt
@@ -1,0 +1,10 @@
+package com.mapbox.vision.sync.syncmanager
+
+internal interface SyncManager {
+
+    fun start()
+
+    fun stop()
+
+    fun syncSessionDir(path: String)
+}
