+++ MapboxVision/src/main/java/com/mapbox/vision/utils/FileUtils.kt
@@ -34,4 +34,10 @@ internal object FileUtils {
 
         return "${getAbsoluteDir(dir.absolutePath)}/"
     }
+
+    fun moveFilesFromOneDirToAnother(fromDirPath: String, toDirPath: String) {
+        val fileFrom = File(fromDirPath)
+        val fileTo = File(toDirPath)
+        fileFrom.renameTo(fileTo)
+    }
 }
