+++ MapboxVision/src/main/java/com/mapbox/vision/video/videoprocessor/VideoProcessorListener.kt
@@ -4,5 +4,5 @@ import com.mapbox.vision.mobile.core.models.VideoClip
 
 internal interface VideoProcessorListener {
 
-    fun onVideoClipsReady(videoClips: HashMap<String, VideoClip>, videoDir: String, jsonFile: String)
+    fun onVideoClipsReady(videoClips: HashMap<String, VideoClip>, videoDir: String, sessionStartMillis: Long)
 }
